Moved utils functions into a new utils.py module.
