+++ tests/conftest.py
@@ -1,3 +1,4 @@
+import os, pathlib
 import sys
 import asyncio
 import logging
